Make sure paths app got is really in mounted volume
Issue description:
SD card volume can be accessed through adb when unmounted.

Steps to reproduce the issue:
1. Insert SD card into DUT and check directory "/storage/SD_CARD_UUID",
where is can be accessed and the contents can be modified;
2. Unmount SD card from Settings;
3. Check this directory again.

Expected Results:
Directory "/storage/SD_CARD_UUID" should not exist.

Actual Results:
This directory still exist and the contents of it can be modified.

Change-Id: Ie15692abdddd010a7dda81034c01ddc4eea6f628
Signed-off-by: Qin Chao <chao.qin@intel.com>