Do not kill attaching process when removing task.
When a process of service is starting for someone binds it,
before attach application to ActivityManagerService, there
is an activity in the same package is removed from recent app.
The service process will be reported as start timeout because
it has not linkToDeath yet. And also the caller to the service
will not receive onServiceDisconnected.

Change-Id: I1e0b50923f5b761ba44267e6e4e932bed3f1847f