Fix the bug in replacing a fragment with itself.
Summary:
We are facing a bug that some fragments in our app got removed
randomly. We debugged the app and it seems there is a bug in
the fragment API.
When we we are trying to replace a fragment with itself. We will
try to make the operation a null op, then both `op.fragment` and
`f` will be reset to null.
op.fragment = f = null;
But since it's inside a loop, if the list `mAdded` still have
items to be visited, the code will continue to the next
iteration. Then it will compare the fragment with null, and
reach the line:
mManager.removeFragment(old, mTransition, mTransitionStyle);
Eventually it will delete all left fragments in the list `mAdded`.
This diff adds a break to abort the searching after we null
the operation.
Test Plan:
1) Tested with our app. No fragment is randomly removed.
Revert Plan:

Change-Id: I5426570440ae5869bdb1d443eda255e36ff9f066