Move service from starting list once onStartCommand get called
Currently bg service will be marked as starting when it get started.
But it will be only get removed from the starting list when it get
binded. This caused the BOOT_COMPLETE intent delay more than 1min when
there are lot of service filter this intent.
Move service from the starting list if onStartCommand is completed.

Change-Id: I5eea83cc13f7b22b474c451e9990f0ad0f2c0698