Fix graphical artifact in the fisheye effect
The fisheye effect generates a graphical artifact close to the center of
the picture, due to bad precision and division by zero in the shader.
The problem is fixed by making a small change in the shader, so that the
picture is uniformly scaled close to the center instead. This avoids the
problem and looks as expected, without affecting the performance.

Change-Id: If69e6e4cdff2ae51e13edc78a979c86630085a9a