[StatusBar] Fix the background of lockscreen showing the art work of removed media notification
Symptom: The background of lockscreen shows the art work of removed media notification

Root cause:
In Android O, StatusBar inflate notification view by AsyncTask(AsyncInflationTask), and the notification entry will only be add to mNotificationData after the inflation is completed.
It makes a timing issue for media notification if following conditions are met:
1. there already exists one none-media notification which is posted by the same app as the incoming media notification.
2. any notification posted and its inflation is completed before the media notification and then triggers the below function.

In such case, the media notification entry hasn't been added to mNotificationData yet so we can't find any media controller in this list(line:2249 ~ 2272).
Then we check mMediaSessionManager and try to find any MediaController that has the same package name in mNotificationData list(line:2273 ~ 2305).
It will find a MediaController which should belong to the media notification, but then be assigned to the none-media notification since both have the same package name.

Change-Id: I80f01f0f50d2a6af76b78eb760b1db25fc618601