FFix service cannot start when service removed from restartList
At 'realStartServiceLocked', 'app.thread.scheduleCreateService' throws
another remote exception, then calls the 'scheduleServiceRestartLocked'.
This time, the process was killed and calls 'killServicesLocked'. The
service can not bringDown, and if parameter 'allowRestart' is 'false',
the service will be removed from restartList. If 'resetRestartCounter'
is not called, the next start service will be returned at
'if (!whileRestarting && r.restartDelay > 0)' every time.

Change-Id: If8643af736c41ef657347c3d8306f4f40db027ca